don't declare local variable "options"
In main.c, the global variable "options" is used all over, except in
main(), which contains a shadowing declaration of it.  Remove this
local variable, and use the global "options" instead.  Also don't
pollute conf.c's namespace with the global variable with a CPP hack
(to be removed when we have fixed more of ncmpc's namespace
pollution).