mpdclient: moved code to playlist.c
Move everything which manipulates the mpdclient_playlist struct to
playlist.c.  Many of the functions get a mpdclient pointer instead of
a playlist; this will be changed later.

The functions mpdclient_playlist_update() and
mpdclient_playlist_update_changes() stay in mpdclient.c for now, since
they are tightly connected to the client code.