screen: don't handle CMD_PLAY
CMD_PLAY is mapped to the Enter key by default.  Some screens
(mis-)use it for navigation, for entering a directory.  The playlist
screen actually plays the file under the cursor.  On all other
screens, pressing Enter may be confusing, since it is more of a
"confirmation" key instead of "play".  We should probaby add
CMD_CONFIRM later, but for now, disable the general CMD_PLAY handler
in screen.c.  This fixes bug 240.