list_window: bell when searches wrap
NCMPC triggers a bell whenever a search wraps. (Note that this means a
double-bell whenever a wrapped search finds nothing.) I can see this
being reasonable, if the playlist is short and the user is aware of
which direction they want to search in. However if you predominantly
use long playlists and don't know which direction your target is in,
this means you get bells on half your searches.

Since I could see people wanting it either way, my best idea is making
the bell optional. Here's a patch to do that, defaulting to the
current behavior. (I'm a little confused about defaults - looks like
there are only default values for some in options.c - but it looks
like this behaves as desired.)