list_window: Added message when selection is enabled/disabled.
The user should get a visual feedback, when he enables or disables the visual selection,
so I added a message which is shown in the status screen. Perhaps there should be a flag
on the right top, when visual selection is enabled to not confuse users, who accidentally
enable it.