list_window: Tweaked visual mode behavior
Tweaked visual mode behavior so that the start of a visual selection is
always one end of it when making large movements (page up/down, H/M/L,
home/end).  To clarify, suppose you start at the middle of the screen,
in visual mode, then:

 (1) get to the bottom with j
 (2) go to the top with page up
 (3) go back to one above middle with j
 (4) j once more

Previously, the resulting selection would be (1) middle to bottom, (2)
top to bottom, (3) one above middle to bottom, and (4) middle only.
With the change, we will get the intuitive (1) middle to bottom, (2) top
to middle, (3) one above middle to middle, and (4) middle only.