Renamed lyrics plugins
Lyrics plugins are tried in alphabetic order by filename.  Numeric
prefixes make this clearer, and make it easier to keep hd.sh first.
I've moved LyricWiki ahead of Leo's Lyrics, mostly because it's easier
for a user to fix incorrect lyrics on a wiki, but also because Leo's
Lyrics appears to return partial matches if no full match is found (e.g.
a song with the same name and completely different artist)