Center the cursor on a second Ctrl-L
To aid in getting to info/lyrics screen, or in selecting ranges
involving the currently playing song.  Ctrl-L followed by M will achieve
the same result as long as the playlist is at least as long as the
screen, but this is faster to type and will work for short playlists
too.

An alternative method would be to center the cursor whenever Ctrl-L will
not cause the screen to move; this could be implemented as an option
perhaps, but would be annoying if you just wanted to repaint the screen.