screen_browser: select/add for visual selection
The select (toggle in playlist) and add (append to playlist) commands now work
for visual selections in browser screens. The status output should eventually
be reconsidered - a message is wastefully printed for each item before it is overwritten by one displaying a count of items added.

The "enter" (CMD_PLAY) command has been disabled for multiple selection for
now; this is better than acting just on the item at the end of the selection
range.  Eventually it should do something fancy like appending all selected
items (directory, song, or playlist) then playing the first song of the first
item.