mpdclient: use the "idle" command to reduce CPU and network usage
This patch makes mpdclient and the rest of ncmpc use the new
mpd_glib_source class.  When mpdclient_get_connection() is called, it
sends the "noidle" command to MPD.  Upon mpdclient_put_connection()
(which is called before returning control back to GLib), mpdclient.c
re-enters "idle" mode.  The "idle" event handler lives in main.c, and
replaces timer_mpd_update().

Currently, the update timer is still used when MPD is playing, to keep
the progress bar and the time up to date.  This is subject to change
soon.

Servers which do not support "idle" are detected automatically.