screen_browser: removed duplicate line clearing
whline(w, ' ', width); was used to clean the rest of the line. The new
function row_paint_text takes care of that already, so we don't have to do
it twice. We had to switch the default case to row_paint_text as well to
cover all possibilities.

This fixes a bug where the last digit of the playtime of the last visible
row in the browse screen was overwritten with a space character.