screen_artist: tracks w/o album tag != all tracks
The `album' variable used to store a pointer to an empty string, which was
somewhere in memory, for the virtual album `All tracks', but this didn't allow
seeing all tracks without an album tag as one album.
Now, the address of the char array `ALL_TRACKS' is assigned to `album', when
`All Tracks' are viewed.