switch from GMainLoop to boost::asio
This adds a rather heavy dependency and bloats the ncmpc binary by 100
kB, but it gives us an easy to use C++ API, and eliminates an
important reason for using GLib.

GLib's I/O event loop was a nice thing when I added it in commit
f0aa88e5c18de0b1e4f5d85dd5c3ae5d9be76bc3 ten years ago.  ncmpc had
been using GLib already, so it was a natural choice without adding
another dependency.  But now that ncmpc is C++, we don't need most of
the GLib toolbox anymore, and there are better I/O event loops.

There's libevent, libev, libuv and others, but they are plain C and
I want ncmpc code to benefit from C++'s safety.

There are few C++ libraries for that, but Boost appears to be the one
which is available everywhere.  Of course, boost::asio has its
downsides; just like most Boost libraries, it suffers from severe
bloat, but inspite of that, it appears to be the best choice for
ncmpc.