Fix config file overriding environment variable
The convention for most programs is that command line options override
environment variables which in turn override config files. Removed the
test for empty as the option should be overridden unconditionally if it
exists in the environment.

In the old implementation, the host was set from the environment
variable in the first pass and then overridden by the config file in the
second pass.