Window: add method GetSize()
This replaces the `size` field.  On ncurses, this adds function call
overhead, but we may eventually be able to eliminate it in optimized
builds using the NCURSES_OPAQUE macro; on the other hand, not
duplicating this piece of information may be a good idea.  The
motivation to do this now is to have a light-weight Window class with
just a pointer.